Successful compilation with Make2
